refactor(ios): inject UserSettingsServicing into WidgetDataSyncService
Couples were hidden: syncAll called UserSettingsService.shared directly
while BudgetService was already injected. Aligns the DI pattern and
unblocks test coverage — the service had zero tests before.

- Extract resolveCurrency(_:) so the policy is testable without network
- Open up init; keep .shared defaults so every caller stays unchanged
- Move MockUserSettingsService to Helpers/ for reuse across suites

Addresses review comment #3 on PR #400.
